+++ astropy/extern/plugins/pytest_remotedata/plugin.py
@@ -16,32 +16,37 @@ def pytest_addoption(parser):
         "--remote-data", nargs="?", const='any', default='none',
         help="run tests with online data")
 
-    parser.addini('remote_data_strict',
+    parser.addini(
+        'remote_data_strict',
         "If 'True', tests will fail if they attempt to access the internet "
         "but are not explicitly marked with 'remote_data'",
         type="bool", default=False)
-
 
 
 def pytest_configure(config):
     config.getini('markers').append(
         'remote_data: Apply to tests that require data from remote servers')
     config.getini('markers').append(
-        'internet_off: Apply to tests that should only run when network access is deactivated')
+        'internet_off: Apply to tests that should only run when '
+        'network access is deactivated')
 
     strict_check = bool(config.getini('remote_data_strict'))
 
     remote_data = config.getoption('remote_data')
-    if remote_data not in ['astropy', 'any', 'none']:
+    if remote_data not in ['astropy', 'any', 'github', 'none']:
         raise pytest.UsageError(
             "'{}' is not a valid source for remote data".format(remote_data))
 
     # Monkeypatch to deny access to remote resources unless explicitly told
     # otherwise
-    if strict_check and remote_data != 'any':
-        turn_off_internet(
-            verbose=config.option.verbose,
-            allow_astropy_data=(remote_data == 'astropy'))
+    if strict_check:
+        if remote_data == 'github':  # GitHub only
+            turn_off_internet(
+                verbose=config.option.verbose, allow_github_data=True)
+        elif remote_data != 'any':  # Astropy + GitHub or nothing
+            turn_off_internet(
+                verbose=config.option.verbose,
+                allow_astropy_data=(remote_data == 'astropy'))
 
 
 def pytest_unconfigure():
@@ -65,13 +70,23 @@ def pytest_runtest_setup(item):
         raise ValueError("remote_data and internet_off are not compatible")
 
     if remote_data is not None:
-        source = remote_data.kwargs.get('source', 'any')
-        if source not in ('astropy', 'any'):
-            raise ValueError("source should be 'astropy' or 'any'")
+        if len(remote_data.args) > 0:
+            source = remote_data.args[0]
+        else:
+            source = remote_data.kwargs.get('source', 'any')
+
+        if source not in ('astropy', 'any', 'github'):
+            raise ValueError("source should be 'astropy', 'any', or 'github'")
 
         if remote_data_config == 'none':
             pytest.skip("need --remote-data option to run")
+        elif remote_data_config == 'github':
+            if source in ('any', 'astropy'):
+                pytest.skip("need --remote-data or "
+                            "--remote-data=astropy option to run")
         elif remote_data_config == 'astropy':
+            # When --remote-data=astropy is given, skip tests simply
+            # marked as --remote-data
             if source == 'any':
                 pytest.skip("need --remote-data option to run")